WineHaus.-basic-apis/products/serializers.py
from rest_framework import serializers
from .models import Product,Category,MadeOf,Country,Brand


class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = '__all__'



class CategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = '__all__' 


class MadeOfSerializer(serializers.ModelSerializer):
    class Meta:
        model = MadeOf
        fields = '__all__'


class CountrySerializer(serializers.ModelSerializer):
    class Meta:
        model = Country
        fields = '__all__'


class BrandSerializer(serializers.ModelSerializer):
    class Meta:
        model = Brand
        fields = '__all__'

