WineHaus.-basic-apis/requirement.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
dj-rest-auth==6.0.0
Django==4.2.16
django-allauth==65.0.2
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
requests==2.32.3
rest-framework-simplejwt==0.0.2
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
